+++ Sources/SwiftMusic/SoundComposition/SoundNode.swift
@@ -1,7 +1,6 @@
 import Foundation
 
 internal indirect enum _SoundNode: Sendable {
-    case group([any Sound])
     case sample(String)
     case fileSample(fileURL: URL, rootPitch: Pitch)
     case sampleBank(SampleBank)